Обернул код в функцию, экспортировал ее в новый модуль
